chore<MatchinPostDto>:두 개로 나눠진 파일 하나로 변경
